Tree window pan and zoom

+++ src/main/java/com/bauproject/pokemontree/Tree.java
@@ -115,7 +115,7 @@ public class Tree {
     }
 
     // Add with image name
-    private Node addRecursive(Node current, Color color, String imgName, ColorEnum type, int index, int depth) {
+    private Node addRecursive(Node current, Color color, String imgName, ColorEnum type, long index, long depth) {
         if (current == null) {
             // Read image files
             return new Node(color, imgName, index, depth);
